refactor interaction
